zdnn: refactor codebase + add docs (#16178)
* zdnn: initial matmul refactor


* ggml-zdnn: rm static from funcs


* ggml-zdnn: update ggml-zdnn.h


* ggml-zdnn: change header files to hpp


* ggml-zdnn: switch to common.hpp


* ggml-zdnn: move mulmat forward around


* ggml-zdnn: rm inline from utils


* ggml-zdnn: code cleanup


* docs: add zDNN docs


---------
