+++ ggml/src/ggml-zdnn/.gitignore
@@ -1,0 +1,1 @@
+zdnn.h
